changed the password constraints
